requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
paho-mqtt==2.1.0
pydantic==2.8.2
python-dotenv==1.0.1
Jinja2==3.1.4
